pyproject.toml
[project]
name = "mtbls-ws3"
version = "v3.0.0"
description = "MetaboLights WebService v3"
authors = [{ name = "MetaboLights Team", email = "metabolights-dev@ebi.ac.uk" }]
requires-python = ">=3.12,<4"
readme = "README.md"
license = "Apache-2.0"
dependencies = [
    "pydantic>=2.10.2,<3",
    "python-dateutil>=2.9.0.post0,<3",
    "pytz~=2024.2",
    "pyyaml>=6.0.2",
    "dependency-injector>=4.42.0,<5",
    "jsonpath-ng>=1.7.0,<2",
    "ujson>=5.10.0,<6",
    "aiocache>=0.12.2",
    "aiofiles>=23.2.1",
    "jsonschema>4.21.1",
    "email-validator>2.1.1",
    "metabolights-utils>=1.4.6,<2",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
ws3 = [
    "psycopg2-binary>=2.9.9,<3",
    "flower>=2.0.1,<3",
    "pymongo>=4.10.1,<5",
    "gunicorn>=23.0.0,<24",
    "uvloop>=0.21.0,<0.22",
    "celery[redis]>=5.4.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.28",
    "asyncpg>=0.29.0",
    "redis>=5.2.1",
    "fastapi>=0.115.5",
    "uvicorn>0.29.0",
    "pyjwt>=2.9.0",
    "asgi-correlation-id>=4.3.4,<5",
    "greenlet>=3.1.1",
]

[dependency-groups]
dev = [
    "import-linter~=2.1",
    "ruff>=0.8.3,<0.9",
    "commitizen>=4.1.0,<5",
    "pre-commit>=4.0.1,<5",
    "mkdocs-material>=9.5.49,<10",

]
test = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0,<0.25",
    "pytest-cov>=6.0.0,<7",
    "aiosqlite>=0.20.0,<0.21",
]

[tool.uv]
default-groups = [
    "dev",
    "test",
]

[tool.hatch.build.targets.wheel]
include = ["mtbls"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
omit = [
  ".*",
  "*/tests/*",
  "*/site-packages/*",
  "*/docs/*",
  "*/site/*",
  "*/dist/*",
  ".*/*",
  "dist*/",
  "validation*/",
  "scripts*/",
]

[tool.coverage.report]

[tool.coverage.html]
directory = ".coverage_html_report"

[tool.commitizen]
name = "cz_conventional_commits"
version = "3.4.0"
tag_format = "v$major.$minor.$patch"
version_files = [

    "pyproject.toml:version",
    "mtbls/__init__.py:__version__",
]
bump_message = "bump: version $current_version -> $new_version"
update_changelog_on_bump = true

[tool.pytest.ini_options]
addopts = "-ra -q -v"
testpaths = [
    "tests"
]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff.lint]
extend-select = ["E4", "E7", "E9", "F", "I", "G", "SLF", "PTH", "Q"]
fixable = ["ALL"]

[tool.importlinter]
include_external_packages = true
root_packages = ["mtbls"]

[[tool.importlinter.contracts]]
name = "Architecture Layer Dependencies"
type = "layers"
layers = [
    "mtbls.run",
    "mtbls.infrastructure | mtbls.presentation",
    "mtbls.application",
    "mtbls.domain"
]

[[tool.importlinter.contracts]]
name = "Restricted Async Task Imports"
type = "forbidden"
source_modules = [

   "mtbls.domain",
   "mtbls.application.services",
   "mtbls.infrastructure",
   "mtbls.presentation.cli",
]
forbidden_modules = [

    "mtbls.application.decorators.async_task",
]
ignore_imports = [
]

[[tool.importlinter.contracts]]
name = "Restricted Dependency Injection"
type = "forbidden"
source_modules = [
   "mtbls",
]
forbidden_modules = [
    "dependency_injector",
]
ignore_imports = [
  "mtbls.application.remote_tasks.** -> dependency_injector",
  "mtbls.presentation.rest_api.core.core_router -> dependency_injector",
  "mtbls.presentation.rest_api.groups.**.routers.** -> dependency_injector",
  "mtbls.run.**.main -> dependency_injector",
  "mtbls.run.**.containers -> dependency_injector",
  "mtbls.run.**.initialization -> dependency_injector",
  "mtbls.run.subscribe -> dependency_injector",
  "mtbls.run.rest_api.submission.base_container -> dependency_injector",
  "mtbls.run.cli.validation.container -> dependency_injector"
]

[[tool.importlinter.contracts]]
name = "Independent Implementations in Infrastructure Layer"
type = "independence"
modules = [
   "mtbls.infrastructure.auth.mtbls_ws2",
   "mtbls.infrastructure.auth.standalone",
   "mtbls.infrastructure.persistence.db.postgresql",
   "mtbls.infrastructure.persistence.db.sqlite",
   "mtbls.infrastructure.cache.in_memory",
   "mtbls.infrastructure.cache.redis",
   "mtbls.infrastructure.cache.redis_sentinel",
   "mtbls.infrastructure.pub_sub.celery",
   "mtbls.infrastructure.pub_sub.threading",
]

ignore_imports = [
  "mtbls.infrastructure.pub_sub.connection.redis -> mtbls.infrastructure.cache.redis.redis_config",
  "mtbls.infrastructure.pub_sub.connection.redis_sentinel -> mtbls.infrastructure.cache.redis_sentinel.redis_sentinel_config",
  "mtbls.infrastructure.cache.redis_sentinel.redis_sentinel_config -> mtbls.infrastructure.cache.redis.redis_config"
]

[[tool.importlinter.contracts]]
name = "Independent Rest API Groups in Presentation Layer"
type = "independence"
modules = [
   "mtbls.presentation.rest_api.groups.auth",
   "mtbls.presentation.rest_api.groups.submission",
   "mtbls.presentation.rest_api.groups.curation",
   "mtbls.presentation.rest_api.groups.public"
]

ignore_imports = [
  "mtbls.presentation.rest_api.groups.*.v1.** -> mtbls.presentation.rest_api.groups.auth.v1.routers.dependencies",
]

[[tool.importlinter.contracts]]
name = "External Framework Independence (Infrastructure)"
type = "forbidden"
source_modules = [
   "mtbls.domain",
   "mtbls.application",
   "mtbls.presentation",
]
forbidden_modules = [
    "redis",
    "sqlalchemy",
    "celery",
    "elasticsearch",
    "psycopg2",
    "uvicorn",
    "flower"
]
ignore_imports = [
]

[[tool.importlinter.contracts]]
name = "External Framework Independence (Presentation)"
type = "forbidden"
source_modules = [
   "mtbls.domain",
   "mtbls.application",
]
forbidden_modules = [
    "fastapi",
    "click",
    "asgi-correlation-id"
]
ignore_imports = [
]

[[tool.importlinter.contracts]]
name = "No Business Logic In Infrastructure Layer"
type = "forbidden"
source_modules = [
   "mtbls.infrastructure",
]
forbidden_modules = [
   "mtbls.application.use_cases",
   "mtbls.application.remote_tasks",
   "mtbls.application.decorators",
   "mtbls.application.context",
]
ignore_imports = [

   "mtbls.** -> mtbls.application.context.request_tracker",
   "mtbls.** -> mtbls.application.context.async_task_registry",
   "mtbls.** -> mtbls.application.decorators.validate"
]

[[tool.importlinter.contracts]]
name = "No Business Logic In Utils"
type = "forbidden"
source_modules = [
   "mtbls.application.utils",
]
forbidden_modules = [
   "mtbls.domain",
   "mtbls.application.use_cases",
   "mtbls.application.remote_tasks",
   "mtbls.application.decorators",
   "mtbls.application.request_tracker",
]
ignore_imports = [
]
